+++ backend/bait-api/routes/api.php
@@ -189,12 +189,12 @@ Route::middleware('auth:api')->prefix('follows')->group(function () {
 });
 
 Route::middleware('auth:api')->prefix('chats')->group(function () {
+    Route::get('/chattable-users', [ChatController::class, 'getChattableUsers']);
     Route::get('/',         [ChatController::class, 'index']);
     Route::post('/',        [ChatController::class, 'store']);
     Route::get('/{chat}',   [ChatController::class, 'show']);
     Route::get('{chat}/messages',  [MessageController::class, 'index']); 
     Route::post('{chat}/messages', [MessageController::class, 'store']); 
-    Route::get('/chattable-users', [ChatController::class, 'getChattableUsers']);
 });
 
 /*end UserInteractions routes*/